+++ CHANGES.rst
@@ -2,10 +2,13 @@ Changelog
 =========
 
 
-1.0.1 (unreleased)
-------------------
+1.1 (unreleased)
+----------------
 
-- Nothing changed yet.
+- Make plugins (by default inactive) ``getGroupMembers`` pluggable, so that
+  integrators may provide its own utility providing new ``IGetGroupMembers``
+  interface promising to return the actual group members.
+  [jensens]
 
 
 1.0 (2016-04-04)